+++ test/SharpLink.IntegrationTests/IntegrationBehaviorTests.cs
@@ -1799,7 +1799,7 @@ public class IntegrationBehaviorTests
             {
                 try
                 {
-                    await server.RunAsync(cts.Token);
+                    await server.RunUntilStoppedAsync(cts.Token);
                 }
                 catch (OperationCanceledException)
                 {