+++ test/SharpLink.IntegrationTests/RpcChannelCallShapeIntegrationTests.cs
@@ -243,7 +243,7 @@ public class RpcChannelCallShapeIntegrationTests
             {
                 try
                 {
-                    await server.RunAsync(cts.Token);
+                    await server.RunUntilStoppedAsync(cts.Token);
                 }
                 catch (OperationCanceledException)
                 {